Show account details
Adds a button to show detailed information of an account. Currently, this is read only but it might be used in the future to change e.g. the name.

In a next step, this should also show the secret such that the information can be used to transfer accounts to a new device. However, I'm unsure how this should be done. So any hints would be welcome.